refactor: Game 생성자 삭제
